+++ packages/design/lib/src/icons.dart
@@ -12,8 +12,10 @@ abstract class IconsWeebi {
       Icon(Icons.format_list_bulleted, color: ColorsWeebi.orangeArticle);
   static const articlePhotos =
       Icon(Icons.image, color: ColorsWeebi.orangeArticle);
-  static const storefront =
-      Icon(Icons.store, color: ColorsWeebi.boutique);
+  static const accounting = FontAwesomeIcons.calculator;
+  static const chartsFinancial = FontAwesomeIcons.chartPie;
+  static const boutique =
+      Icon(Icons.storefront, color: ColorsWeebi.boutique);
   static const boutiqueIconData = Icons.storefront;
   static const contacts = Icon(Icons.contacts, color: ColorsWeebi.blueContact);
   static const contactsIconData = Icons.contacts;