+++ webapp/lib/views/screens/crud_detail_screen.dart
@@ -1,10 +1,10 @@
-import 'package:awesome_dialog/awesome_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:go_router/go_router.dart';
 import 'package:web_admin/app_router.dart';
 import 'package:web_admin/generated/l10n.dart';
+import 'package:web_admin/utils/app_dialogs.dart';
 import 'package:web_admin/utils/app_focus_helper.dart';
 import 'package:web_admin/views/widgets/card_elements.dart';
 import 'package:web_admin/views/widgets/portal_master_layout/portal_master_layout.dart';
@@ -50,29 +50,25 @@ class _CrudDetailScreenState extends State<CrudDetailScreen> {
 
       final lang = Lang.of(context);
 
-      final dialog = AwesomeDialog(
+      AppDialog.show(
         context: context,
-        dialogType: DialogType.question,
+        dialogType: AppDialogType.question,
         title: lang.confirmSubmitRecord,
         width: kDialogWidth,
         btnOkText: lang.yes,
         btnOkOnPress: () {
-          final d = AwesomeDialog(
+          AppDialog.show(
             context: context,
-            dialogType: DialogType.success,
+            dialogType: AppDialogType.success,
             title: lang.recordSubmittedSuccessfully,
             width: kDialogWidth,
             btnOkText: 'OK',
             btnOkOnPress: () => GoRouter.of(context).go(RouteUri.crud),
           );
-
-          d.show();
         },
         btnCancelText: lang.cancel,
         btnCancelOnPress: () {},
       );
-
-      dialog.show();
     }
   }
 
@@ -81,29 +77,25 @@ class _CrudDetailScreenState extends State<CrudDetailScreen> {
 
     final lang = Lang.of(context);
 
-    final dialog = AwesomeDialog(
+    AppDialog.show(
       context: context,
-      dialogType: DialogType.infoReverse,
+      dialogType: AppDialogType.warning,
       title: lang.confirmDeleteRecord,
       width: kDialogWidth,
       btnOkText: lang.yes,
       btnOkOnPress: () {
-        final d = AwesomeDialog(
+        AppDialog.show(
           context: context,
-          dialogType: DialogType.success,
+          dialogType: AppDialogType.success,
           title: lang.recordDeletedSuccessfully,
           width: kDialogWidth,
           btnOkText: 'OK',
           btnOkOnPress: () => GoRouter.of(context).go(RouteUri.crud),
         );
-
-        d.show();
       },
       btnCancelText: lang.cancel,
       btnCancelOnPress: () {},
     );
-
-    dialog.show();
   }
 
   @override